PLT-525 - Delegated Access - review fix

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppspersononprobationuserapi/data/DelegatedAccess.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.hmppspersononprobationuserapi.data
 
 data class DelegatedAccess(
-  var initiatedUserId: Int,
-  var delegatedUserId: Int,
+  var initiatedUserId: Long,
+  var delegatedUserId: Long,
 )
